+++ controller/index.js
@@ -181,38 +181,35 @@ var Controller = function (agol, BaseController) {
     var options = Utils.createExportOptions(req, info)
     var dirName = path.dirname(options.output)
     var fileName = path.basename(options.output)
-    var awsPath = options.output
-    // Make sure we can still find files that were created under the old cache-keying regime
-    if (info.version !== 3) {
-      var key = Utils.createCacheKey(req.params, req.query)
-      fileName = fileName.replace(/\/full\//, '/' + key + '/')
-      awsPath = path.join(dirName, fileName)
-    }
-
     agol.files.exists(dirName, fileName, function (exists) {
       var exportStatus = Utils.determineStatus(req, info)
       if (exists) {
+        if (exportStatus === info.retrieved_at) return controller._returnFile(req, res, options.output, info)
         // check to see if the file itself is Expired
         // if it is, we will still serve it but we need to enqueue a new one
-        if (exportStatus !== info.retrieved_at) {
-          agol.generateExport(options, function (err) {
-            if (err) agol.log.error(err)
-          })
-        }
-        return controller._returnFile(req, res, awsPath, info)
-      }
-      // if a job is already running or we don't actually have the data in the cache yet
-      // hand off to returnStatus for a 202 response
-      var jobInProgress = ['queued', 'start', 'progress', 'fail'].indexOf(exportStatus) > -1
-      var processing = info.status === 'Processing'
-      var error = exportStatus === 'fail' ? new Error('Export process failed') : undefined
-      if (error) error.code = 500
-      if (jobInProgress || processing) return controller._returnStatus(req, res, info, error)
+        agol.generateExport(options, function (err, status) {
+          if (err) agol.log.error(err)
+          // always serve filtered data from the same cache as the full export
+          if (info.status === 'Cached' && (req.query.where || req.query.geometry)) {
+            controller._returnStatus(req, res, info, error)
+          } else {
+            controller._returnFile(req, res, options.output, info)
+          }
+        })
+      } else {
+        // if a job is already running or we don't actually have the data in the cache yet
+        // hand off to returnStatus for a 202 response
+        var jobInProgress = ['queued', 'start', 'progress', 'fail'].indexOf(exportStatus) > -1
+        var processing = info.status === 'Processing'
+        var error = exportStatus === 'fail' ? new Error('Export process failed') : undefined
+        if (error) error.code = 500
+        if (jobInProgress || processing) return controller._returnStatus(req, res, info, error)
 
-      // only enqueue a job if it's not already queued or running
-      agol.generateExport(options, function (err, status, created) {
-        controller._returnStatus(req, res, status, err)
-      })
+        // only enqueue a job if it's not already queued or running
+        agol.generateExport(options, function (err, status, created) {
+          controller._returnStatus(req, res, status, err)
+        })
+      }
     })
   }